refactor(stores): decompose useMapStore into domain slices (P2-2)
The store declared 39 state fields and 47 actions in a single 590-line
module, so any change to one domain touched the same file as every other.
Split the implementation into six Zustand slice factories (map instance,
route, search, transit, a11y, sheet) over a shared MapStore type, and keep
useMapStore.ts as the aggregator re-exporting the same public surface.

getMapPadding now delegates to computeMapPadding(sidebarCollapsed) so the
sheet slice can compute padding without importing the store back.
